refactor: MyFunding 패키지 -> Member 패키지로 변경(MyFunding 도메인 없어서 별도의 패키지 어색)
