+++ src/main/java/com/theocean/fundering/domain/member/controller/MemberController.java
@@ -4,10 +4,17 @@ import com.theocean.fundering.domain.member.dto.EmailRequestDTO;
 import com.theocean.fundering.domain.member.dto.MemberSettingRequestDTO;
 import com.theocean.fundering.domain.member.service.MemberService;
 import com.theocean.fundering.domain.member.dto.MemberSignUpRequestDTO;
+import com.theocean.fundering.domain.member.dto.MyFundingHostResponseDTO;
+import com.theocean.fundering.domain.member.dto.MyFundingManagerResponseDTO;
+import com.theocean.fundering.domain.member.dto.MyFundingSupporterResponseDTO;
+import com.theocean.fundering.domain.member.service.MyFundingService;
+import com.theocean.fundering.global.dto.PageResponse;
 import com.theocean.fundering.global.jwt.userInfo.CustomUserDetails;
 import com.theocean.fundering.global.utils.ApiUtils;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.web.PageableDefault;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -65,5 +72,5 @@ public class MemberController {
     public ResponseEntity<?> member(){
         return ResponseEntity.ok().body(ApiUtils.success(null));
     }
-
+    
 }
